+++ applications/serial_lte_modem/src/slm_ppp.h
@@ -8,9 +8,6 @@
 
 #include <stdbool.h>
 
-/* Whether to forward CGEV notifications to the SLM UART. */
-extern bool slm_fwd_cgev_notifs;
-
 /** @retval 0 on success. */
 int slm_ppp_init(void);
